New approach to ranking skills
The rankings are so wishy washy, so let's simplify it to a word map!
This should hopefully scale better, though unfortunately this JS library
has a major flaw in that it cannot guarantee it can fit everything! And
then things get left out entirely if they don't fit :(.

Safest thing to do is to prevent rotation, avoid randomizing the
placement, and give it some extra space that we'll trim off. That
annoyingly can mess with scrolling, since the word cloud is async, but
it's not that bad and likely won't be noticed often because this takes
place further down the page.